% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_containment_health}
\alias{calculate_containment_health}
\title{Calculate containment and health index}
\usage{
calculate_containment_health(df)
}
\arguments{
\item{df}{A data.frame produced by a call to \code{\link[=calculate_subindices]{calculate_subindices()}}.}
}
\value{
A numeric value for mean index scores for
\strong{containment and health index}.
}
\description{
Calculate containment and health index
}
\examples{
x <- get_data_actions(json = get_json_actions(ccode = "AFG",
                                              on = "2020-09-01"))
y <- calculate_subindices(df = x$policyActions)

calculate_containment_health(df = y)

}
\author{
Ernest Guevarra
}
