% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_economic_support}
\alias{calculate_economic_support}
\title{Calculate economic support index}
\usage{
calculate_economic_support(df)
}
\arguments{
\item{df}{A data.frame produced by a call to \code{\link[=calculate_subindices]{calculate_subindices()}}.}
}
\value{
A numeric value for mean index scores for \strong{economic support index}.
}
\description{
Calculate economic support index
}
\examples{
x <- get_data_actions(json = get_json_actions(ccode = "AFG",
                                              on = "2020-09-01"))
y <- calculate_subindices(df = x$policyActions)

calculate_economic_support(df = y)

}
\author{
Ernest Guevarra
}
