% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_indices}
\alias{calculate_indices}
\title{Calculate OxCGRT index or indices}
\usage{
calculate_indices(df)
}
\arguments{
\item{df}{A data.frame produced by a call to \code{\link[=calculate_subindices]{calculate_subindices()}}.}
}
\value{
A tibble calculated OxCGRT indices
}
\description{
Calculate OxCGRT index or indices
}
\examples{
##
x <- get_data_actions(json = get_json_actions(ccode = "AFG",
                                              on = "2020-09-01"))

y <- calculate_subindices(df = x$policyActions)

calculate_indices(df = y)

}
\author{
Ernest Guevarra
}
