% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_stringency}
\alias{calculate_stringency}
\title{Calculate OxCGRT stringency index}
\usage{
calculate_stringency(df)
}
\arguments{
\item{df}{A data.frame produced by a call to \code{\link[=calculate_subindices]{calculate_subindices()}}.}
}
\value{
A numeric value for mean index scores for \strong{stringency index}.
}
\description{
Calculate OxCGRT stringency index
}
\examples{
x <- get_data(json = get_json_actions(ccode = "AFG",
                                      from = NULL,
                                      to = "2020-09-01"))
y <- calculate_subindices(df = x$policyActions)

calculate_stringency(df = y)

}
\author{
Ernest Guevarra
}
