% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_subindex}
\alias{calculate_subindex}
\title{Calculate OxCGRT sub-index score for a single indicator}
\usage{
calculate_subindex(indicator_code, value, flag_value)
}
\arguments{
\item{indicator_code}{Two-character code for indicators of policy measures
tracked by \strong{Oxford COVID-19 Government Response Tracker}.}

\item{value}{An integer for recorded policy value on the ordinal scale for
given policy type.}

\item{flag_value}{Either logical or binary value (0, 1) for recorded flag
value. If NA, corresponds to policy type with no flag.}
}
\value{
A numeric value between 0 to 100.
}
\description{
Calculate OxCGRT sub-index score for a single indicator
}
\examples{
calculate_subindex(indicator_code = indicatorData$indicator[1],
                   value = indicatorData$value[1],
                   flag_value = indicatorData$flag_value[1])

}
\author{
\strong{Ernest Guevarra} based on calculation methods by \emph{Hale, Thomas,}
\emph{Noam Angrist, Emily Cameron-Blake, Laura Hallas, Beatriz Kira,}
\emph{Saptarshi Majumdar, Anna Petherick, Toby Phillips, Helen Tatlow,}
\emph{Samuel Webster (2020). Oxford COVID-19 Government Response Tracker,}
\emph{Blavatnik School of Government.}
}
