% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_subindices}
\alias{calculate_subindices}
\title{Calculate sub-index scores for all indicators}
\usage{
calculate_subindices(
  df,
  indicator_code = "policy_type_code",
  value = "policyvalue_actual",
  flag_value = "flagged",
  add = TRUE
)
}
\arguments{
\item{df}{A data.frame containing per indicator values required for
calculating sub-index scores. This data.frame will be something like the
policy actions data.frame produced by a call to \code{\link[=get_data_actions]{get_data_actions()}}.}

\item{indicator_code}{A character value specifying the name of the column
in \code{df} containing the indicator codes.}

\item{value}{A character value specifying the name of the column in \code{df}
containing the values assigned to each indicator}

\item{flag_value}{A character value specifying the name of the column in \code{df}
containing the flag values for each indicator}

\item{add}{Logical. Should sub-indices for each indicator be added to \code{df}?
Default is TRUE.}
}
\value{
If \code{add} is TRUE (default), returns a tibble composed of the input
data.frame \code{x} with an added column named score for the calculated
sub-indices. If \code{add} is FALSE, returns a tibble of 2 columns with the
first column for the indicator codes and the second column named score for
the calculated sub-indices.
}
\description{
Calculate sub-index scores for all indicators
}
\examples{
x <- get_data_actions(json = get_json_actions(ccode = "AFG",
                                              on = "2020-09-01"))
calculate_subindices(df = x$policyActions)

}
\author{
Ernest Guevarra
}
