% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Codebook for the Oxford COVID-19 Government Response Tracker}
\format{
A tibble with 28 rows and 6 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{ID} \tab Policy indicator identifier \cr
   \code{Name} \tab Name of policy indicator \cr
   \code{Description} \tab Description of policy indicator \cr
   \code{Measurement} \tab Measurement of policy indicator \cr
   \code{Coding} \tab Coding of measurement \cr
   \verb{Policy Group} \tab Name of group policy indicator \cr
}
}
\source{
https://github.com/OxCGRT/covid-policy-tracker/blob/master/documentation/codebook.md
}
\usage{
codebook
}
\description{
Codebook for the Oxford COVID-19 Government Response Tracker
}
\keyword{datasets}
