% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get policy actions and stringency data from JSON}
\usage{
get_data(json)
}
\arguments{
\item{json}{A JSON string, URL or file created using \code{\link[=get_json_actions]{get_json_actions()}} or
a vector of JSON strings or URLs.}
}
\value{
A named list of two tibbles. The first tibble is named
\code{policyActions}. The second tibble is named \code{stringencyData}
}
\description{
Get policy actions and stringency data from JSON
}
\examples{
## Get relevant JSON
x <- get_json_actions(ccode = "AFG", from = NULL, to = "2020-07-16")

## Get data on policy actions and stringency index
get_data(x)

}
