% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_v1.R
\name{get_data_v1}
\alias{get_data_v1}
\alias{get_data_v1_complete}
\alias{get_data_v1_timeseries}
\title{Get legacy data version 1}
\usage{
get_data_v1(
  type = c("complete", "timeseries"),
  pcode = c("s1", "s2", "s3", "s4", "s5", "s6", "s7", "s11", "s12", "s13"),
  notes = TRUE,
  directory = NULL,
  tidy = FALSE
)

get_data_v1_complete(notes = TRUE, directory = NULL)

get_data_v1_timeseries(
  pcode = c("s1", "s2", "s3", "s4", "s5", "s6", "s7", "s11", "s12", "s13"),
  directory = NULL,
  tidy = FALSE
)
}
\arguments{
\item{type}{A character value for type of dataset to retrieve. Either
\emph{"complete"} for the full dataset or \emph{"timeseries"} for the disaggregated
policy indicators datasets. Default is \emph{"complete"}.}

\item{pcode}{A character value for policy indicator code to retrieve if
\code{type} is \emph{"timeseries"}. Ignored if \code{type} is \emph{"complete"}. Can be one of
\emph{"s1"}, \emph{"s2"}, \emph{"s3"}, \emph{"s4"}, \emph{"s5"}, \emph{"s6"}, \emph{"s7"}, \emph{"s8"}, \emph{"s9"},
\emph{"s10"}, \emph{"s11"}, \emph{"s12"}, or \emph{"s13"}.}

\item{notes}{Logical. Should \emph{"complete"} dataset be retrieved with notes or
not. Default is TRUE. Ignored if \code{type} is \emph{"timeseries"}.}

\item{directory}{If NULL (default), required dataset is read into current
environment. If specified, required dataset is downloaded into this
directory with the filename given by the OxCGRT.}

\item{tidy}{Logical. Should \emph{"timeseries"} dataset be converted to tidy
format? Default is FALSE.}
}
\value{
If \code{directory} is NULL, a tibble (or a list of two tibbles) of the
specified dataset. Otherwise, a file path/s to the downloaded file/s.
}
\description{
Get legacy data version 1
}
\examples{
get_data_v1_complete()
get_data_v1_timeseries()
get_data_v1()

}
