% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json_time}
\alias{get_json_time}
\title{Get JSON for stringency index data over time}
\usage{
get_json_time(from = "2020-01-02", to = as.character(Sys.Date()))
}
\arguments{
\item{from}{Start date for stringency index data to be collected. This can go
as far back as 2020-01-02 (Default). Format <YYYY-MM-DD>.}

\item{to}{End data for stringency index data to be collected. This defaults
to current date. Format <YYYY-MM-DD>}
}
\value{
A character object for specified JSON endpoint
}
\description{
Get JSON for stringency index data over time
}
\examples{
get_json_time()

}
