% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorData}
\alias{indicatorData}
\title{Example indicator data for sub-index calculations}
\format{
A tibble with 14 rows and 6 columns\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{indicator} \tab Policy indicator code \cr
   \code{value} \tab Policy indicator value \cr
   \code{flag_value} \tab Policy indicator flag value \cr
   \code{max_value} \tab Maximum value for policy indicator \cr
   \code{flag} \tab Does the policy indicator have a flag value? 1 = Yes; 0 = No \cr
   \code{score} \tab Policy indicator score from 0 - 100 \cr
}
}
\source{
https://github.com/OxCGRT/covid-policy-tracker/blob/master/documentation/index_methodology.md
}
\usage{
indicatorData
}
\description{
Example indicator data for sub-index calculations
}
\keyword{datasets}
