% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxcgrt_data_files}
\alias{oxcgrt_data_files}
\title{Table of OxCGRT dataset files available from OxCGRT GitHub data repositories}
\format{
A tibble with 90 rows and 6 columns\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{version} \tab OxCGRT dataset version. Either \strong{v1}, \strong{v2}, or \strong{final} \cr
   \code{filename} \tab Filename of dataset file \cr
   \code{path} \tab Path of dataset file \cr
   \code{url} \tab GitHub URL to dataset file \cr
   \code{url_raw} \tab GitHub raw URL to dataset file \cr
   \code{sha} \tab SHA-1 hash value for dataset file \cr
}
}
\usage{
oxcgrt_data_files
}
\description{
Table of OxCGRT dataset files available from OxCGRT GitHub data repositories
}
\keyword{datasets}
