% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_gov_response}
\alias{calculate_gov_response}
\title{Calculate OxCGRT government response index}
\usage{
calculate_gov_response(df)
}
\arguments{
\item{df}{A data.frame produced by a call to \code{\link[=calculate_subindices]{calculate_subindices()}}.}
}
\value{
A numeric value for mean index scores for
\strong{government response index}.
}
\description{
Calculate OxCGRT government response index
}
\examples{
x <- get_data_actions(json = get_json_actions(ccode = "AFG",
                                              on = "2020-09-01"))
y <- calculate_subindices(df = x$policyActions)

calculate_gov_response(df = y)

}
\author{
Ernest Guevarra
}
