% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_subindex}
\alias{calculate_subindex}
\title{Calculate sub-index score for a single indicator}
\usage{
calculate_subindex(indicator_code, value, flag_value)
}
\arguments{
\item{indicator_code}{Two-character code for indicators of policy measures
tracked by Oxford COVID-19 Government Response Tracker.}

\item{value}{An integer for recorded policy value on the ordinal scale for
given indicator.}

\item{flag_value}{Either logical or binary value (0, 1) for recorded flag
value. If NA, corresponds to indicator with no flag.}
}
\value{
A numeric value between 0 to 100.
}
\description{
Calculate sub-index score for a single indicator
}
\author{
Ernest Guevarra based on calculation methods by Hale, Thomas,
Noam Angrist, Emily Cameron-Blake, Laura Hallas, Beatriz Kira,
Saptarshi Majumdar, Anna Petherick, Toby Phillips, Helen Tatlow,
Samuel Webster (2020). Oxford COVID-19 Government Response Tracker,
Blavatnik School of Government.
}
