% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json_actions}
\alias{get_json_actions}
\title{Get JSON for policy actions data for a specific country for a specific date}
\usage{
get_json_actions(ccode, on = as.character(Sys.Date()))
}
\arguments{
\item{ccode}{ISO 3166-1 alpha-3 country code}

\item{on}{Date on which policy actions data is required. This defaults
to current date. Format \code{YYYY-MM-DD}.}
}
\value{
A character object for specified JSON endpoint
}
\description{
Get JSON for policy actions data for a specific country for a specific date
}
\examples{
get_json_actions(ccode = "AFG", on = as.character(Sys.Date() - 7))

}
