% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcgrt.R
\docType{package}
\name{oxcgrt}
\alias{oxcgrt}
\alias{oxcgrt-package}
\title{An Interface to the Oxford COVID-19 Government Response Tracker API}
\description{
The \strong{Oxford COVID-19 Government Response Tracker (OxCGRT)} tracks and
compares worldwide government responses to the COVID-19 pandemic rigorously
and consistently. \strong{OxCGRT} makes available systematic information in a
consistent way, aiding those who require information have access to it
efficiently for their purposes. This package facilitates access to the
\strong{OxCGRT} data via its API for R users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://como-ph.github.io/oxcgrt/}
  \item \url{https://github.com/como-ph/oxcgrt}
  \item Report bugs at \url{https://github.com/como-ph/oxcgrt/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
